V3::Public specs, add some specs for this
Notably, these specs helped identify an issue where a V3::Public it was
possible to initialize the class with a key that was not on secp384r1. The
class now protects against this by copying the input key into an empty
EC container that has its group pre-initialized to secp384r1. This raises
an error when copying the public key, or alternatively raises on a #check_key
call when the resulting point is not on the group curve.