CI, fix reek plugin in CC (#186)
This plugin has been failing for years because Code Climate insists on running reek under 2.6 with parser 3.1.

Today, I found https://github.com/troessner/reek/pull/1694, which shows an undocumented way to set a target version for reek. This commit sets a target version of 3.1 and fixes some of the "new" findings.